+++ extensions/ql-vscode/src/extension.ts
@@ -15,8 +15,6 @@ import { arch, homedir, platform } from "os";
 import { ensureDir } from "fs-extra";
 import { join } from "path";
 import { dirSync } from "tmp-promise";
-import type { TestHub } from "vscode-test-adapter-api";
-import { testExplorerExtensionId } from "vscode-test-adapter-api";
 import { lt, parse } from "semver";
 import { watch } from "chokidar";
 import {
@@ -25,17 +23,18 @@ import {
 } from "./common/vscode/archive-filesystem-provider";
 import { CliVersionConstraint, CodeQLCliServer } from "./codeql-cli/cli";
 import {
+  ADD_DATABASE_SOURCE_TO_WORKSPACE_SETTING,
+  addDatabaseSourceToWorkspace,
   CliConfigListener,
   DistributionConfigListener,
   GitHubDatabaseConfigListener,
-  isCanary,
   joinOrderWarningThreshold,
   QueryHistoryConfigListener,
   QueryServerConfigListener,
 } from "./config";
 import {
   AstViewer,
-  createIDEServer,
+  createLanguageClient,
   getQueryEditorCommands,
   install,
   TemplatePrintAstProvider,
@@ -85,13 +84,11 @@ import {
 import type { ProgressReporter } from "./common/logging/vscode";
 import {
   extLogger,
-  ideServerLogger,
+  languageServerLogger,
   queryServerLogger,
 } from "./common/logging/vscode";
 import { QueryHistoryManager } from "./query-history/query-history-manager";
 import type { CompletedLocalQueryInfo } from "./query-results";
-import { QLTestAdapterFactory } from "./query-testing/test-adapter";
-import { TestUIService } from "./query-testing/test-ui";
 import { CompareView } from "./compare/compare-view";
 import {
   initializeTelemetry,
@@ -130,7 +127,6 @@ import { DebuggerUI } from "./debugger/debugger-ui";
 import { ModelEditorModule } from "./model-editor/model-editor-module";
 import { TestManager } from "./query-testing/test-manager";
 import { TestRunner } from "./query-testing/test-runner";
-import type { TestManagerBase } from "./query-testing/test-manager-base";
 import { QueryRunner, QueryServerClient } from "./query-server";
 import { QueriesModule } from "./queries-panel/queries-module";
 import { OpenReferencedFileCodeLensProvider } from "./local-queries/open-referenced-file-code-lens-provider";
@@ -175,7 +171,7 @@ function getCommands(
   app: App,
   cliServer: CodeQLCliServer,
   queryRunner: QueryRunner,
-  ideServer: LanguageClient,
+  languageClient: LanguageClient,
 ): BaseCommands {
   const getCliVersion = async () => {
     try {
@@ -193,10 +189,10 @@ function getCommands(
         await Promise.all([
           queryRunner.restartQueryServer(progress, token),
           async () => {
-            if (ideServer.isRunning()) {
-              await ideServer.restart();
+            if (languageClient.isRunning()) {
+              await languageClient.restart();
             } else {
-              await ideServer.start();
+              await languageClient.start();
             }
           },
         ]);
@@ -308,6 +304,14 @@ const codeQlVersionRange = DEFAULT_DISTRIBUTION_VERSION_RANGE;
 // before silently being refused to upgrade.
 const MIN_VERSION = "1.82.0";
 
+function sendConfigTelemetryData() {
+  const config: Record<string, string> = {};
+  config[ADD_DATABASE_SOURCE_TO_WORKSPACE_SETTING.qualifiedName] =
+    addDatabaseSourceToWorkspace().toString();
+
+  telemetryListener?.sendConfigInformation(config);
+}
+
 /**
  * Returns the CodeQLExtensionInterface, or an empty object if the interface is not
  * available after activation is complete. This will happen if there is no cli
@@ -334,6 +338,8 @@ export async function activate(
   install();
 
   const app = new ExtensionApp(ctx);
+
+  sendConfigTelemetryData();
 
   const quickEvalCodeLensProvider = new QuickEvalCodeLensProvider();
   languages.registerCodeLensProvider(
@@ -936,7 +942,7 @@ async function activateWithInstalledDistribution(
   ctx.subscriptions.push(tmpDirDisposal);
 
   void extLogger.log("Initializing CodeQL language server.");
-  const ideServer = createIDEServer(qlConfigurationListener);
+  const languageClient = createLanguageClient(qlConfigurationListener);
 
   const localQueries = new LocalQueries(
     app,
@@ -977,27 +983,8 @@ async function activateWithInstalledDistribution(
   const testRunner = new TestRunner(dbm, cliServer);
   ctx.subscriptions.push(testRunner);
 
-  let testManager: TestManagerBase | undefined = undefined;
-  if (isCanary()) {
-    testManager = new TestManager(app, testRunner, cliServer);
-    ctx.subscriptions.push(testManager);
-  } else {
-    const testExplorerExtension = extensions.getExtension<TestHub>(
-      testExplorerExtensionId,
-    );
-    if (testExplorerExtension) {
-      const testHub = testExplorerExtension.exports;
-      const testAdapterFactory = new QLTestAdapterFactory(
-        testHub,
-        testRunner,
-        cliServer,
-      );
-      ctx.subscriptions.push(testAdapterFactory);
-
-      testManager = new TestUIService(app, testHub);
-      ctx.subscriptions.push(testManager);
-    }
-  }
+  const testManager = new TestManager(app, testRunner, cliServer);
+  ctx.subscriptions.push(testManager);
 
   const testUiCommands = testManager?.getCommands() ?? {};
 
@@ -1021,7 +1008,7 @@ async function activateWithInstalledDistribution(
   void extLogger.log("Registering top-level command palette commands.");
 
   const allCommands: AllExtensionCommands = {
-    ...getCommands(app, cliServer, qs, ideServer),
+    ...getCommands(app, cliServer, qs, languageClient),
     ...getQueryEditorCommands({
       commandManager: app.commands,
       queryRunner: qs,
@@ -1068,21 +1055,21 @@ async function activateWithInstalledDistribution(
   }
 
   void extLogger.log("Starting language server.");
-  await ideServer.start();
+  await languageClient.start();
   ctx.subscriptions.push({
     dispose: () => {
-      void ideServer.stop();
+      void languageClient.stop();
     },
   });
 
-  // Handle visibility changes in the ideserver
+  // Handle visibility changes in the CodeQL language client.
   if (await cliServer.cliConstraints.supportsVisibilityNotifications()) {
     Window.onDidChangeVisibleTextEditors((editors) => {
-      ideServer.notifyVisibilityChange(editors);
+      languageClient.notifyVisibilityChange(editors);
     });
     // Send an inital notification to the language server
     // to set the initial state of the visible editors.
-    ideServer.notifyVisibilityChange(Window.visibleTextEditors);
+    languageClient.notifyVisibilityChange(Window.visibleTextEditors);
   }
 
   // Jump-to-definition and find-references
@@ -1264,7 +1251,7 @@ function getContextStoragePath(ctx: ExtensionContext) {
 async function initializeLogging(ctx: ExtensionContext): Promise<void> {
   ctx.subscriptions.push(extLogger);
   ctx.subscriptions.push(queryServerLogger);
-  ctx.subscriptions.push(ideServerLogger);
+  ctx.subscriptions.push(languageServerLogger);
 }
 
 const checkForUpdatesCommand: keyof PreActivationCommands =