+++ extensions/ql-vscode/src/view/model-editor/__tests__/MethodRow.spec.tsx
@@ -30,6 +30,7 @@ describe(MethodRow.name, () => {
     provenance: "manual",
   };
   const onChange = jest.fn();
+  const onMethodClick = jest.fn();
 
   const viewState = createMockModelEditorViewState();
 
@@ -40,10 +41,12 @@ describe(MethodRow.name, () => {
         methodCanBeModeled={true}
         modeledMethods={[modeledMethod]}
         methodIsUnsaved={false}
+        methodIsSelected={false}
         modelingInProgress={false}
         revealedMethodSignature={null}
         viewState={viewState}
         onChange={onChange}
+        onMethodClick={onMethodClick}
         {...props}
       />,
     );