+++ extensions/ql-vscode/src/view/model-editor/__tests__/ModeledMethodsList.spec.tsx
@@ -33,6 +33,7 @@ describe(ModeledMethodsList.name, () => {
     methodParameters: "(String)",
   });
   const onChange = jest.fn();
+  const onMethodClick = jest.fn();
   const onSaveModelClick = jest.fn();
   const onGenerateFromLlmClick = jest.fn();
   const onStopGenerateFromLlmClick = jest.fn();
@@ -56,11 +57,13 @@ describe(ModeledMethodsList.name, () => {
           ],
         }}
         modifiedSignatures={new Set([method1.signature])}
+        selectedSignatures={new Set()}
         inProgressMethods={new Set()}
         viewState={viewState}
         hideModeledMethods={false}
         revealedMethodSignature={null}
         onChange={onChange}
+        onMethodClick={onMethodClick}
         onSaveModelClick={onSaveModelClick}
         onGenerateFromLlmClick={onGenerateFromLlmClick}
         onStopGenerateFromLlmClick={onStopGenerateFromLlmClick}